automation/requirements.txt
# Build tools (must be first)
setuptools>=68.0.0
wheel>=0.41.0

# Core dependencies
requests>=2.31.0
pandas>=2.0.2
numpy>=2.0.0
python-dotenv>=1.0.0

# Google APIs
google-api-python-client==2.88.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
gspread==5.9.0

# Email
# smtplib is a built-in Python module, no installation needed

# Automation
schedule==1.2.0
APScheduler==3.10.1

# Logging
loguru==0.7.0

# Data processing
openpyxl==3.1.2
xlsxwriter==3.1.2

# HTTP client
httpx==0.24.1

# API Framework
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Development
black==23.3.0
isort==5.12.0
flake8==6.0.0
